requirements.txt
streamlit==1.38.0
streamlit-webrtc==0.47.7
tensorflow-cpu==2.15.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
pandas==2.2.2
plotly==5.21.0
aiortc==1.10.0
av==13.0.0
Pillow==10.4.0
kaleido==1.1.0